+++ Structs/Util.h
@@ -28,10 +28,10 @@ struct size2 {
 size2 makeSize2(int height, int width);
 
 struct bounds {
-  int min, max;
+  float min, max;
 };
 
-bounds makeBounds(int min, int max);
+bounds makeBounds(float min, float max);
 
 //taken from https://www.fluentcpp.com/2017/04/21/how-to-split-a-string-in-c/
 std::vector<std::string> split(const std::string& s, char delimiter);
